requirements.txt
beautifulsoup4>=4.7.1
bs4>=0.0.1
certifi>=2018.11.29
chardet>=3.0.4
Click>=7.0
Flask>=1.0.2
Flask-SQLAlchemy>=2.3.2
fuzzywuzzy>=0.17.0
grapheme>=0.4.0
idna>=2.8
itsdangerous>=1.1.0
Jinja2>=2.10
lxml>=4.3.0
MarkupSafe>=1.1.0
mutagen>=1.42.0
python-Levenshtein>=0.12.0
pytz>=2018.9
PyYAML>=3.13
requests>=2.21.0
six>=1.12.0
soupsieve>=1.7.2
SQLAlchemy>=1.2.16
tzlocal>=1.5.1
urllib3>=1.24.1
wcwidth>=0.1.7
Werkzeug>=0.14.1
wrapt>=1.11.1
